UPSTREAM: fuse: always allow query of st_dev
Fuse mounts without "allow_other" are off-limits to all non-owners.  Yet it
makes sense to allow querying st_dev on the root, since this value is
provided by the kernel, not the userspace filesystem.

Allow statx(2) with a zero request mask to succeed on a fuse mounts for all
users.

Reported-by: Nikolaus Rath <Nikolaus@rath.org>
Change-Id: Ie7215acbfa5ee7cbfe512b3e12a00d1045f8fbc3
Signed-off-by: Miklos Szeredi <mszeredi@redhat.com>